Display hotel overview on manager view
