+++ src/index.html
@@ -31,16 +31,16 @@
         <section class="overview-default" id="hotel-overview-info"></section>
       </article>
     </section>
-      <section class="current-bookings hidden">
-        <h5 class="search-title">Bookings for --</h5>
-        <div class="grid-container" class="grid-header">
-          <div class="grid-item">Date</div>
-          <div class="grid-item">Guest Name</div>
-          <div class="grid-item">Room Type</div>
-          <div class="grid-item">Room #</div>
-          <div class="grid-item">Bed Size</div>
-          <div class="grid-item">Cost</div>
-          <div class="grid-item">Confirmation</div>
+      <section class="today-hotel-bookings hidden">
+        <h5 class="today-hotel-bookings-title">Bookings for --</h5>
+        <div class="grid-container">
+          <div class="grid-header"><u>Date</u></div>
+          <div class="grid-header"><u>Guest Name</u></div>
+          <div class="grid-header"><u>Room Type</u></div>
+          <div class="grid-header"><u>Room #</u></div>
+          <div class="grid-header"><u>Bed Size</u></div>
+          <div class="grid-header"><u>Cost</u></div>
+          <div class="grid-header"><u>Confirmation</u></div>
         </div>
       </section>
       <section class="customer-search-bar hidden">
